requirements.txt
Flask==3.0.3
pdf2docx==0.5.6
Pillow==10.4.0
reportlab==4.2.2
gunicorn==22.0.0
Werkzeug==3.0.3